backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.1.1
django-allauth==0.57.0
requests==2.31.0
pyotp==2.9.0
bleach==6.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
